feat: nebula 2.0 (#263)
* feat: nebula 2.0 & add loading effect when login

* feat: optimize the disconnection logic

* feat: add timeout setting

* feat: fix schema 2.0

* feat: support explore in 2.0

* feat: fix index query bug

* feat: reset state after logout

* fix: filter empty stringVid when explore

* feat: update version

* feat: remove code

+++ app/assets/utils/constant.ts
@@ -1,5 +1,5 @@
 export const enumOfCompare = {
-  int: [
+  int64: [
     {
       label: '==',
       value: '==',
